change: 안마의자·자습 신청 로직을 feature 훅으로 분리
