+++ src/entities/dormitory/api/getDormitory.ts
@@ -8,6 +8,10 @@ import type {
   CleaningZoneDetail 
 } from '@/entities/dormitory/model/dormitory';
 
+function extractList<T>(response: T[] | { data?: T[] }): T[] {
+  return Array.isArray(response) ? response : (response.data ?? []);
+}
+
 type DormitoryMusicResponseItem = Omit<DormitoryMusic, "isLiked"> & {
   isLiked?: boolean;
 };
@@ -20,9 +24,7 @@ export async function getDormitoryMusic(date?: string): Promise<DormitoryMusic[]
     params: date ? { date } : undefined,
   });
 
-  const musicList = Array.isArray(data) ? data : data.data ?? [];
-
-  return musicList.map((music) => ({
+  return extractList(data).map((music) => ({
     ...music,
     isLiked: music.isLiked ?? false,
   }));
@@ -32,12 +34,12 @@ type DormitoryStudentResponse = DormitoryStudent[] | { data?: DormitoryStudent[]
 
 export async function getMassageApplicants(): Promise<DormitoryStudent[]> {
   const { data } = await instance.get<DormitoryStudentResponse>('/dormitory/massages');
-  return Array.isArray(data) ? data : (data.data ?? []);
+  return extractList(data);
 }
 
 export async function getSelfStudyApplicants(): Promise<DormitoryStudent[]> {
   const { data } = await instance.get<DormitoryStudentResponse>('/dormitory/studies');
-  return Array.isArray(data) ? data : (data.data ?? []);
+  return extractList(data);
 }
 
 export async function getMyPenalties(): Promise<MyPenaltyResponse> {
