+++ src/features/massage-chair/model/useApplyMassage.ts
@@ -1,0 +1,36 @@
+"use client";
+
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+import axios, { HttpStatusCode } from "axios";
+import { toast } from "sonner";
+import {
+  dormitoryQueries,
+  dormitoryMutations,
+} from "@/entities/dormitory/api/dormitoryQueries";
+
+export function useApplyMassage() {
+  const queryClient = useQueryClient();
+  const massageQuery = dormitoryQueries.massage();
+
+  return useMutation({
+    mutationFn: dormitoryMutations.applyMassage,
+    onSuccess: () =>
+      queryClient.invalidateQueries({ queryKey: massageQuery.queryKey }),
+    onError: (error) => {
+      const status = axios.isAxiosError(error)
+        ? error.response?.status
+        : undefined;
+
+      if (status === HttpStatusCode.BadRequest) {
+        toast.error("마사지 신청 시간이 아닙니다.");
+        return;
+      }
+      if (status === HttpStatusCode.Conflict) {
+        toast.error("이미 안마의자를 신청했습니다.");
+        return;
+      }
+
+      toast.error("안마의자 신청에 실패했습니다.");
+    },
+  });
+}
