+++ src/widgets/main/ui/MassageApplyCard.tsx
@@ -1,43 +1,17 @@
 "use client";
 
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import axios, { HttpStatusCode } from "axios";
-import { toast } from "sonner";
+import { useQuery } from "@tanstack/react-query";
 import ChairIcon from "@/shared/asset/svg/Chair";
+import { dormitoryQueries } from "@/entities/dormitory/api/dormitoryQueries";
+import { useApplyMassage } from "@/features/massage-chair/model/useApplyMassage";
 import ApplyCard from "./ApplyCard";
-import {
-  dormitoryQueries,
-  dormitoryMutations,
-} from "@/entities/dormitory/api/dormitoryQueries";
 
 const MASSAGE_MAX = 5;
 
 export default function MassageApplyCard() {
-  const queryClient = useQueryClient();
   const massageQuery = dormitoryQueries.massage();
   const { data: applicants = [] } = useQuery(massageQuery);
-
-  const applyMutation = useMutation({
-    mutationFn: dormitoryMutations.applyMassage,
-    onSuccess: () =>
-      queryClient.invalidateQueries({ queryKey: massageQuery.queryKey }),
-    onError: (error) => {
-      const status = axios.isAxiosError(error)
-        ? error.response?.status
-        : undefined;
-
-      if (status === HttpStatusCode.BadRequest) {
-        toast.error("마사지 신청 시간이 아닙니다.");
-        return;
-      }
-      if (status === HttpStatusCode.Conflict) {
-        toast.error("이미 안마의자를 신청했습니다.");
-        return;
-      }
-
-      toast.error("안마의자 신청에 실패했습니다.");
-    },
-  });
+  const applyMutation = useApplyMassage();
 
   return (
     <ApplyCard
