+++ src/widgets/main/ui/StudyApplyCard.tsx
@@ -1,47 +1,17 @@
 "use client";
 
-import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
-import axios, { HttpStatusCode } from "axios";
-import { toast } from "sonner";
+import { useQuery } from "@tanstack/react-query";
 import BookIcon from "@/shared/asset/svg/ApplyStudy";
+import { dormitoryQueries } from "@/entities/dormitory/api/dormitoryQueries";
+import { useApplyStudy } from "@/features/self-study/model/useApplyStudy";
 import ApplyCard from "./ApplyCard";
-import {
-  dormitoryQueries,
-  dormitoryMutations,
-} from "@/entities/dormitory/api/dormitoryQueries";
 
 const STUDY_MAX = 50;
 
 export default function StudyApplyCard() {
-  const queryClient = useQueryClient();
   const studyQuery = dormitoryQueries.study();
   const { data: students = [] } = useQuery(studyQuery);
-
-  const applyMutation = useMutation({
-    mutationFn: dormitoryMutations.applyStudy,
-    onSuccess: () =>
-      queryClient.invalidateQueries({ queryKey: studyQuery.queryKey }),
-    onError: (error) => {
-      const status = axios.isAxiosError(error)
-        ? error.response?.status
-        : undefined;
-
-      if (status === HttpStatusCode.BadRequest) {
-        toast.error("자습 신청 시간이 아닙니다.");
-        return;
-      }
-      if (status === HttpStatusCode.Forbidden) {
-        toast.error("자습 금지 상태입니다.");
-        return;
-      }
-      if (status === HttpStatusCode.Conflict) {
-        toast.error("이미 자습을 신청했습니다.");
-        return;
-      }
-
-      toast.error("자습 신청에 실패했습니다.");
-    },
-  });
+  const applyMutation = useApplyStudy();
 
   return (
     <ApplyCard
